Prevent CSRF attack on course unreg

+++ modules/unreguser/unregcours.php
@@ -40,6 +40,10 @@ if(!isset($_GET['cid']))
 
 if (!isset($doit) or $doit != "yes") {
 
+  if(isset($_SESSION['token'])) {
+      $CSRFToken = $_SESSION['token'];
+  }
+
   $tool_content .= "
     <table width='40%'>
     <tbody>
@@ -48,7 +52,7 @@ if (!isset($doit) or $doit != "yes") {
       	<p>$langConfirmUnregCours:</p><p> <em>".course_code_to_title($cid)."</em>&nbsp;? </p>
 	<ul class='listBullet'>
 	<li>$langYes: 
-	<a href='$_SERVER[PHP_SELF]?u=$uid&amp;cid=$cid&amp;doit=yes' class=mainpage>$langUnregCours</a>
+	<a href='$_SERVER[PHP_SELF]?u=$uid&amp;cid=$cid&amp;doit=yes&CSRFToken=$CSRFToken' class=mainpage>$langUnregCours</a>
 	</li>
 	<li>$langNo: <a href='../../index.php' class=mainpage>$langBack</a>
 	</li></ul>
@@ -58,14 +62,17 @@ if (!isset($doit) or $doit != "yes") {
     </table>";
 
 } else {
-if (isset($uid) and $uid==$_SESSION['uid']) {
+    if (isset($uid) and $uid==$_SESSION['uid']) {
+        if(isset($CSRFToken) and $CSRFToken == $_SESSION['token']) {
             db_query("DELETE from cours_user WHERE cours_id = (SELECT cours_id FROM cours WHERE code = " . quote($cid) . ") AND user_id='$uid'");
-                if (mysql_affected_rows() > 0) {
-                        $tool_content .= "<p class='success_small'>$langCoursDelSuccess</p>";
-                } else {
-                        $tool_content .= "<p class='caution_small'>$langCoursError</p>";
-                }
-         }
+            if (mysql_affected_rows() > 0) {
+                $tool_content .= "<p class='success_small'>$langCoursDelSuccess</p>";
+            } else {
+                $tool_content .= "<p class='caution_small'>$langCoursError</p>";
+            }
+
+        }
+    }
         $tool_content .= "<br><br><div align=right><a href='../../index.php' class=mainpage>$langBack</a></div>";
 }
